backend/requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
fastapi_jwt_auth==0.5.0
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==1.0.0
cryptography==3.4.7
sqlalchemy==1.4.17
aiosqlite==0.17.0
aioredis==1.3.1